Refactored http joiners and injectors
